Adding async support to the long haul test (#457)
