+++ Tests/LongHaulTestApp/GameLogic/Game.cpp
@@ -19,112 +19,6 @@ std::mutex Game::m_displayEventQueueLock;
 #define TEST_DELAY_FAST 2
 
 
-class win32_handle
-{
-public:
-    win32_handle() : m_handle(nullptr)
-    {
-    }
-
-    ~win32_handle()
-    {
-        if (m_handle != nullptr) CloseHandle(m_handle);
-        m_handle = nullptr;
-    }
-
-    void set(HANDLE handle)
-    {
-        m_handle = handle;
-    }
-
-    HANDLE get() { return m_handle; }
-
-private:
-    HANDLE m_handle;
-};
-
-win32_handle g_stopRequestedHandle;
-win32_handle g_workReadyHandle;
-win32_handle g_completionReadyHandle;
-
-void CALLBACK HandleAsyncQueueCallback(
-    _In_ void* context,
-    _In_ async_queue_handle_t queue,
-    _In_ AsyncQueueCallbackType type
-)
-{
-    UNREFERENCED_PARAMETER(context);
-    UNREFERENCED_PARAMETER(queue);
-
-    switch (type)
-    {
-    case AsyncQueueCallbackType::AsyncQueueCallbackType_Work:
-        SetEvent(g_workReadyHandle.get());
-        break;
-
-    case AsyncQueueCallbackType::AsyncQueueCallbackType_Completion:
-        SetEvent(g_completionReadyHandle.get());
-        break;
-    }
-}
-
-
-DWORD WINAPI background_thread_proc(LPVOID lpParam)
-{
-    UNREFERENCED_PARAMETER(lpParam);
-    HANDLE hEvents[3] =
-    {
-        g_workReadyHandle.get(),
-        g_completionReadyHandle.get(),
-        g_stopRequestedHandle.get()
-    };
-
-    async_queue_handle_t queue;
-    uint32_t sharedAsyncQueueId = 0;
-    CreateSharedAsyncQueue(
-        sharedAsyncQueueId,
-        AsyncQueueDispatchMode::AsyncQueueDispatchMode_Manual,
-        AsyncQueueDispatchMode::AsyncQueueDispatchMode_Manual,
-        &queue);
-
-    bool stop = false;
-    while (!stop)
-    {
-        DWORD dwResult = WaitForMultipleObjectsEx(3, hEvents, false, INFINITE, false);
-        switch (dwResult)
-        {
-        case WAIT_OBJECT_0: // work ready
-            DispatchAsyncQueue(queue, AsyncQueueCallbackType_Work, 0);
-
-            if (!IsAsyncQueueEmpty(queue, AsyncQueueCallbackType_Work))
-            {
-                // If there's more pending work, then set the event to process them
-                SetEvent(g_workReadyHandle.get());
-            }
-            break;
-
-        case WAIT_OBJECT_0 + 1: // completed 
-                                // Typically completions should be dispatched on the game thread, but
-                                // for this simple XAML app we're doing it here
-            DispatchAsyncQueue(queue, AsyncQueueCallbackType_Completion, 0);
-
-            if (!IsAsyncQueueEmpty(queue, AsyncQueueCallbackType_Completion))
-            {
-                // If there's more pending completions, then set the event to process them
-                SetEvent(g_completionReadyHandle.get());
-            }
-            break;
-
-        default:
-            stop = true;
-            break;
-        }
-    }
-
-    CloseAsyncQueue(queue);
-    return 0;
-}
-
 // Loads and initializes application assets when the application is loaded.
 Game::Game(const std::shared_ptr<DX::DeviceResources>& deviceResources) :
     m_deviceResources(deviceResources),
@@ -138,21 +32,9 @@ Game::Game(const std::shared_ptr<DX::DeviceResources>& deviceResources) :
     // Register to be notified if the Device is lost or recreated
     m_deviceResources->RegisterDeviceNotify(this);
     m_sceneRenderer = std::unique_ptr<Renderer>(new Renderer(m_deviceResources));
-
-    g_stopRequestedHandle.set(CreateEvent(nullptr, true, false, nullptr));
-    g_workReadyHandle.set(CreateEvent(nullptr, false, false, nullptr));
-    g_completionReadyHandle.set(CreateEvent(nullptr, false, false, nullptr));
-
+    
     XblInitialize();
-    uint32_t sharedAsyncQueueId = 0;
-    CreateSharedAsyncQueue(
-        sharedAsyncQueueId,
-        AsyncQueueDispatchMode::AsyncQueueDispatchMode_Manual,
-        AsyncQueueDispatchMode::AsyncQueueDispatchMode_Manual,
-        &m_queue);
-    AddAsyncQueueCallbackSubmitted(m_queue, nullptr, HandleAsyncQueueCallback, &m_callbackToken);
-
-    m_hBackgroundThread = CreateThread(nullptr, 0, background_thread_proc, nullptr, 0, nullptr);
+    InitializeAsync();
 
     XblUserCreateHandle(&m_user);
 }
@@ -160,17 +42,8 @@ Game::Game(const std::shared_ptr<DX::DeviceResources>& deviceResources) :
 void Game::RegisterInputKeys()
 {
     m_input->RegisterKey(Windows::System::VirtualKey::S, ButtonPress::SignIn);
-    m_input->RegisterKey(Windows::System::VirtualKey::P, ButtonPress::GetUserProfile);
-    m_input->RegisterKey(Windows::System::VirtualKey::F, ButtonPress::GetFriends);
     m_input->RegisterKey(Windows::System::VirtualKey::T, ButtonPress::StartTests);
     m_input->RegisterKey(Windows::System::VirtualKey::R, ButtonPress::StartTestsFast);
-    m_input->RegisterKey(Windows::System::VirtualKey::A, ButtonPress::GetAchievementsForTitle);
-    m_input->RegisterKey(Windows::System::VirtualKey::Number1, ButtonPress::ToggleSocialGroup1);
-    m_input->RegisterKey(Windows::System::VirtualKey::Number2, ButtonPress::ToggleSocialGroup2);
-    m_input->RegisterKey(Windows::System::VirtualKey::Number3, ButtonPress::ToggleSocialGroup3);
-    m_input->RegisterKey(Windows::System::VirtualKey::Number4, ButtonPress::ToggleSocialGroup4);
-    m_input->RegisterKey(Windows::System::VirtualKey::Number5, ButtonPress::ToggleSocialGroup5);
-    m_input->RegisterKey(Windows::System::VirtualKey::C, ButtonPress::ImportCustomList);
 }
 
 Game::~Game()
@@ -270,18 +143,6 @@ void Game::OnGameUpdate()
                 InitializeTestFramework();
                 m_testDelay = TEST_DELAY_FAST;
                 m_testsStarted = true;
-            }
-
-            try
-            {
-                if (m_testsStarted)
-                {
-                    HandleTests();
-                }
-            }
-            catch (const std::exception& e)
-            {
-                Log("[Error] " + std::string(e.what()));
             }
         }
     }
