+++ Tests/LongHaulTestApp/LongHaulTestApp.vcxproj.filters
@@ -69,6 +69,9 @@
     <ClCompile Include="SocialIntegrations.cpp">
       <Filter>Integrations</Filter>
     </ClCompile>
+    <ClCompile Include="AsyncIntegration.cpp">
+      <Filter>Integrations</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="App.h" />
